// dashboard/src/utils/hasura-api/generated/schemas/dropComputedFieldArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { QualifiedTable } from './qualifiedTable';

export interface DropComputedFieldArgs {
  table: QualifiedTable;
  /** Name of the computed field */
  name: string;
  /** When set to true, all the dependent items on this computed field are also dropped */
  cascade?: boolean;
  /** Name of the source database of the table */
  source?: string;
}